+++ clientScripts/loader.js
@@ -162,13 +162,13 @@
 		})
 
 		System.import('es' + (legacy ? 5 : 6) + '/main')
-	}
 
-	if ('serviceWorker' in navigator) {
-		navigator.serviceWorker
-			.register("/worker.js")
-			.catch(function (err) {
-				throw err
-			})
+		if ('serviceWorker' in navigator) {
+			navigator.serviceWorker
+				.register("/worker.js")
+				.catch(function (err) {
+					throw err
+				})
+		}
 	}
 })()