+++ client/page/board.ts
@@ -1,3 +1,3 @@
-// Format a board name and title into cannoncal board header format
+// Format a board name and title into cannonical board header format
 export const formatHeader = (name: string, title: string): string =>
 	`/${name}/ - ${title}`