+++ server/websockets/accounts_test.go
@@ -26,7 +26,7 @@ func (*DB) TestRegistrationValidations(c *C) {
 	}{
 		{"", "123456", idTooShort},
 		{r21, "123456", idTooLong},
-		{"123", "12345", passwordTooShort},
+		{"123", "", passwordTooShort},
 		{"123", r31, passwordTooLong},
 	}