+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Project uses [Semantic Versioning](http://semver.org/)
 ##vNext
 * Change:
 	- Required Go version to 1.7
+* Fix:
+	- Client HTTPS detection
 
 ##v2.1.0-alpha - 2016-07-04
 * License: