+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Project uses [Semantic Versioning](http://semver.org/)
 * Change:
 	- Required Go version to 1.7
 	- Move origin configuration to server flag
+	- Restrict board names to [a-z0-9]{1,3}
 * Fix:
 	- Client HTTPS detection
 	- SystemJS not loading in Palemoon